+++ include/pkgi.h
@@ -5,7 +5,7 @@
 
 #include <dbglogger.h>
 
-#define PKGI_UPDATE_URL     "http://update.pkgi.tk/version.txt" //api.github.com/repos/bucanero/pkgi-ps3/releases/latest
+#define PKGI_UPDATE_URL     "https://api.github.com/repos/bucanero/pkgi-ps3/releases/latest"
 #define PKGI_VERSION        "1.0.8"
 
 // values compatible with psp2/ctrl.h header
@@ -31,6 +31,8 @@
 #define PKGI_PKG_FOLDER "/dev_hdd0/packages"
 #define PKGI_RAP_FOLDER "/dev_hdd0/exdata"
 #define PKGI_TMP_FOLDER "/dev_hdd0/tmp"
+#define PKGI_QUEUE_FOLDER "/dev_hdd0/vsh/task"
+#define PKGI_INSTALL_FOLDER "/dev_hdd0/vsh/game_pkg"
 
 typedef struct pkgi_input {
     uint64_t delta;   // microseconds from previous frame